+++ activejob/test/adapters/backburner.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-require "support/backburner/inline"
-
-ActiveJob::Base.queue_adapter = :backburner