+++ activejob/CHANGELOG.md
@@ -1,3 +1,8 @@
+*   Remove `BackburnerAdapter` from `ActiveJob`.
+    `ActiveJob::QueueAdapters::BackburnerAdapter` is implemented inside the latest `backburner` gem release.
+
+    *Dino Maric*, *Nathan Esquenazi* 
+
 *   Remove `sucker_punch` as an adapter option [since author himself recommends using AJ's own AsyncAdapter](https://github.com/brandonhilkert/sucker_punch?tab=readme-ov-file#faq).
     If you're using this adapter, change to `adapter: async` for the same functionality.